adding simple examples for reference
