+++ db/scratch/random.js
